# Codechange: Scan station catchment tiles when removing station from nearby towns/industries.