@JGRennison 2024-02-19
Players can (and often do) bulldoze whole towns. If a town has no house tiles it wouldn't be found and the reference would be left dangling.