@PeterN 2024-02-19
`ClearTownHouse()` removes the station from `Town::stations_near` via `RemoveNearbyStations()`, so I don't think that would be affected.